refactor: helper method to print non-repeatable warnings
